+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/state/TxState.java
@@ -108,7 +108,7 @@ public interface TxState
 
     public interface Visitor
     {
-        void visitNodeLabelChanges( long id, Set<Long> added, Set<Long> removed );
+        void visitNodeLabelChanges( long id, Set<Integer> added, Set<Integer> removed );
 
         void visitAddedIndex( IndexDescriptor element, boolean isConstraintIndex );
 
@@ -128,9 +128,9 @@ public interface TxState
 
     public abstract Iterable<NodeState> nodeStates();
 
-    public abstract DiffSets<Long> labelStateNodeDiffSets( long labelId );
+    public abstract DiffSets<Long> labelStateNodeDiffSets( int labelId );
 
-    public abstract DiffSets<Long> nodeStateLabelDiffSets( long nodeId );
+    public abstract DiffSets<Integer> nodeStateLabelDiffSets( long nodeId );
 
     public abstract DiffSets<DefinedProperty> nodePropertyDiffSets( long nodeId );
 
@@ -141,12 +141,12 @@ public interface TxState
     /**
      * Returns all nodes that, in this tx, have had labelId added.
      */
-    public abstract Set<Long> nodesWithLabelAdded( long labelId );
+    public abstract Set<Long> nodesWithLabelAdded( int labelId );
 
     /**
      * Returns all nodes that, in this tx, have had labelId removed.
      */
-    public abstract DiffSets<Long> nodesWithLabelChanged( long labelId );
+    public abstract DiffSets<Long> nodesWithLabelChanged( int labelId );
 
     // Temporary: Should become DiffSets<Long> of all node changes, not just deletions
     public abstract DiffSets<Long> nodesDeletedInTx();
@@ -155,15 +155,15 @@ public interface TxState
 
     public abstract boolean nodeIsDeletedInThisTx( long nodeId );
 
-    public abstract DiffSets<Long> nodesWithChangedProperty( long propertyKeyId, Object value );
+    public abstract DiffSets<Long> nodesWithChangedProperty( int propertyKeyId, Object value );
 
-    public abstract Map<Long, Object> nodesWithChangedProperty( long propertyKeyId );
+    public abstract Map<Long, Object> nodesWithChangedProperty( int propertyKeyId );
 
     public abstract boolean relationshipIsAddedInThisTx( long relationshipId );
 
     public abstract boolean relationshipIsDeletedInThisTx( long relationshipId );
 
-    public abstract UpdateTriState labelState( long nodeId, long labelId );
+    public abstract UpdateTriState labelState( long nodeId, int labelId );
 
     public abstract void relationshipDoDelete( long relationshipId );
 
@@ -182,17 +182,17 @@ public interface TxState
 
     public abstract void graphDoRemoveProperty( Property removedProperty );
 
-    public abstract void nodeDoAddLabel( long labelId, long nodeId );
+    public abstract void nodeDoAddLabel( int labelId, long nodeId );
 
-    public abstract void nodeDoRemoveLabel( long labelId, long nodeId );
+    public abstract void nodeDoRemoveLabel( int labelId, long nodeId );
 
 
 
     // SCHEMA RELATED
 
-    public abstract DiffSets<IndexDescriptor> indexDiffSetsByLabel( long labelId );
+    public abstract DiffSets<IndexDescriptor> indexDiffSetsByLabel( int labelId );
 
-    public abstract DiffSets<IndexDescriptor> constraintIndexDiffSetsByLabel( long labelId );
+    public abstract DiffSets<IndexDescriptor> constraintIndexDiffSetsByLabel( int labelId );
 
     public abstract DiffSets<IndexDescriptor> indexChanges();
 
@@ -202,10 +202,10 @@ public interface TxState
 
     public abstract DiffSets<UniquenessConstraint> constraintsChanges();
 
-    public abstract DiffSets<UniquenessConstraint> constraintsChangesForLabel( long labelId );
+    public abstract DiffSets<UniquenessConstraint> constraintsChangesForLabel( int labelId );
 
-    public abstract DiffSets<UniquenessConstraint> constraintsChangesForLabelAndProperty( long labelId,
-                                                                                          long propertyKey );
+    public abstract DiffSets<UniquenessConstraint> constraintsChangesForLabelAndProperty( int labelId,
+                                                                                          int propertyKey );
 
     public abstract void indexRuleDoAdd( IndexDescriptor descriptor );
