+++ community/kernel/src/test/java/org/neo4j/kernel/impl/api/integrationtest/NodeGetUniqueFromIndexLookupIT.java
@@ -45,7 +45,7 @@ import static org.junit.Assert.assertTrue;
 
 public class NodeGetUniqueFromIndexLookupIT extends KernelIntegrationTest
 {
-    private long labelId, propertyKeyId;
+    private int labelId, propertyKeyId;
 
     @Before
     public void createKeys() throws Exception
